# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CaribeTour API",
        "description": "API documentation for CaribeTour project",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/statuses": {
            "get": {
                "summary": "Get list of statuses",
                "description": "Display a listing of the resource.",
                "operationId": "642664c912621220ca61f5b175605ca1",
                "parameters": [
                    {
                        "name": "filter",
                        "in": "query",
                        "description": "Filter statuses by fields (e.g., name)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limit the number of statuses returned",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "summary": "Create a new status",
                "description": "Store a newly created resource in storage.",
                "operationId": "a59343c409c50ebd023a31575312c3cf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Status Name"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Status created successfully"
                    }
                }
            }
        },
        "/api/v1/statuses/{id}": {
            "get": {
                "summary": "Get a status by ID",
                "description": "Display the specified resource.",
                "operationId": "82cfdc00fe3517e81f5a1dcbbc856034",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "put": {
                "summary": "Update a status",
                "description": "Update the specified resource in storage.",
                "operationId": "999e68ef5de3a83d92ff3ef44e3c904b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Status Name"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status updated successfully"
                    }
                }
            },
            "delete": {
                "summary": "Delete a status",
                "description": "Remove the specified resource from storage.",
                "operationId": "6d73d030aadbeef48b57629b93df4987",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Status deleted successfully"
                    }
                }
            }
        },
        "/api/v1/terminals": {
            "get": {
                "summary": "Get list of terminals",
                "description": "Display a listing of the resource.",
                "operationId": "bebbac2893d672f93a64e5857ffdd4c3",
                "parameters": [
                    {
                        "name": "filter",
                        "in": "query",
                        "description": "Filter terminals by fields (e.g., name)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limit the number of terminals returned",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "summary": "Create a new terminal",
                "description": "Store a newly created resource in storage.",
                "operationId": "91047f26a45793d9844b358f7b3018bc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Terminal Name"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "123 Main St"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Terminal created successfully"
                    }
                }
            }
        },
        "/api/v1/terminals/{id}": {
            "get": {
                "summary": "Get a terminal by ID",
                "description": "Display the specified resource.",
                "operationId": "3b4eb45ec7a5f7f05cf6c2e777d0cdc0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "put": {
                "summary": "Update a terminal",
                "description": "Update the specified resource in storage.",
                "operationId": "5dccba6c4a91055c0601a5b10f81f78c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Terminal Name"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "456 Elm St"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Terminal updated successfully"
                    }
                }
            },
            "delete": {
                "summary": "Delete a terminal",
                "description": "Remove the specified resource from storage.",
                "operationId": "6366519f374551b50d58b4e41a8edd6a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Terminal deleted successfully"
                    }
                }
            }
        },
...
    }
}